refactor: move BrowserNotifyModal component
