+++ web/components/modals/BrowserNotifyModal/BrowserNotifyModal.module.scss
